LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-overlay-expected.txt


FAIL dialog.show() should not put popover dialog in top layer assert_equals: Computed overlay expected (string) "none" but got (undefined) undefined
FAIL dialog.showModal() should put popover dialog in top layer assert_equals: Computed overlay expected (string) "none" but got (undefined) undefined
FAIL dialog.showPopover() should put popover dialog in top layer assert_equals: Computed overlay expected (string) "none" but got (undefined) undefined
FAIL div.showPopover() should put popover div in top layer assert_equals: Computed overlay expected (string) "none" but got (undefined) undefined
